DL-17926: Removed hip feature flag and unused NPS related code
