+++ app/config/ServiceBindings.scala
@@ -20,7 +20,7 @@ import auth.{AuthClientConnector, DefaultAuthClientConnector}
 import connectors.{CitizenDetailsConnector, DefaultCitizenDetailsConnector, DefaultNpsConnector, NpsConnector}
 import play.api.{Configuration, Environment}
 import play.api.inject.{Binding, Module}
-import services.{DefaultProtectionService, ProtectionService}
+import services.HipProtectionService
 
 class ServiceBindings extends Module {
 
@@ -28,7 +28,7 @@ class ServiceBindings extends Module {
     bindServices() ++ bindConnectors()
 
   private def bindServices(): Seq[Binding[_]] = Seq(
-    play.api.inject.bind(classOf[ProtectionService]).to(classOf[DefaultProtectionService]).eagerly()
+    play.api.inject.bind(classOf[HipProtectionService]).toSelf.eagerly()
   )
 
   private def bindConnectors(): Seq[Binding[_]] = Seq(
