--- app/events/NPSCreateLTAEvent.scala
@@ -1,38 +1,0 @@
-/*
- * Copyright 2024 HM Revenue & Customs
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package events
-
-import play.api.libs.json.JsObject
-import uk.gov.hmrc.http.HeaderCarrier
-
-class NPSCreateLTAEvent(
-    nino: String,
-    npsRequestBodyJs: JsObject,
-    npsResponseBodyJs: JsObject,
-    statusCode: Int,
-    path: String
-)(implicit hc: HeaderCarrier)
-    extends NPSBaseLTAEvent(
-      ltaAuditType = "CreateAllowance",
-      transactionName = "create-pensions-lifetime-allowance",
-      nino = nino,
-      npsRequestBodyJs = npsRequestBodyJs,
-      npsResponseBodyJs = npsResponseBodyJs,
-      statusCode = statusCode,
-      path = path,
-      extraDetail = Map()
-    )
