--- app/model/PensionDebit.scala
@@ -1,35 +1,0 @@
-/*
- * Copyright 2024 HM Revenue & Customs
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package model
-
-import play.api.libs.json._
-
-case class PensionDebit(startDate: String, amount: Double)
-
-object PensionDebit {
-
-  val jsonWrites: Writes[PensionDebit] = new Writes[PensionDebit] {
-    override def writes(o: PensionDebit): JsValue = Json.obj(
-      "pensionDebitStartDate"     -> o.startDate,
-      "pensionDebitEnteredAmount" -> o.amount
-    )
-  }
-
-  val jsonReads: Reads[PensionDebit] = Json.reads
-
-  implicit val pdFormat: Format[PensionDebit] = Format[PensionDebit](jsonReads, jsonWrites)
-}
