--- app/model/ReadProtectionsModel.scala
@@ -1,46 +1,0 @@
-/*
- * Copyright 2024 HM Revenue & Customs
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package model
-
-import play.api.libs.functional.syntax._
-import play.api.libs.json._
-
-case class ReadProtectionsModel(
-    nino: String,
-    psaCheckReference: String,
-    lifetimeAllowanceProtections: Option[List[ReadProtection]] = None
-)
-
-object ReadProtectionsModel {
-
-  implicit val formats: Format[ReadProtectionsModel] = {
-
-    val jsonReads: Reads[ReadProtectionsModel] = new Reads[ReadProtectionsModel] {
-      override def reads(json: JsValue): JsResult[ReadProtectionsModel] =
-        (JsPath \ "nino")
-          .read[String]
-          .and((JsPath \ "pensionSchemeAdministratorCheckReference").read[String])
-          .and((JsPath \ "protections").readNullable[List[ReadProtection]])(ReadProtectionsModel.apply _)
-          .reads(json)
-    }
-
-    val jsonWrites: Writes[ReadProtectionsModel] = Json.writes[ReadProtectionsModel]
-
-    Format(jsonReads, jsonWrites)
-  }
-
-}
